switched to floats

+++ DrawingLibrary/ExtensionMethods.cs
@@ -55,9 +55,9 @@ namespace DrawingLibrary
         {
             uint argb = (uint)color.ToArgb();
             return new Vector3(
-                (double)(argb >> 16 & 255) / 255,
-                (double)(argb >> 8 & 255) / 255,
-                (double)(argb & 255) / 255);
+                (float)(argb >> 16 & 255) / 255,
+                (float)(argb >> 8 & 255) / 255,
+                (float)(argb & 255) / 255);
         }
         public static Color ToColor(this Vector3 vector3)
         {
